[#64] SplashView 이동
