+++ Soomsil-USaint/Application/Feature/Splash/View/SplashView.swift
@@ -1,0 +1,16 @@
+//
+//  SplashView.swift
+//  Soomsil-USaint
+//
+//  Created by 이조은 on 1/6/25.
+//
+
+import SwiftUI
+
+struct SplashView: View {
+    var body: some View {
+        Image("splash")
+            .resizable()
+            .aspectRatio(contentMode: .fill)
+    }
+}
